+++ domain/src/main/java/ru/dpopkov/knownics/domain/AbstractTranslation.java
@@ -4,17 +4,37 @@ import lombok.Getter;
 import lombok.Setter;
 import ru.dpopkov.knownics.domain.user.AppUser;
 
+import javax.persistence.*;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
+
 @Getter
 @Setter
+@Embeddable
+@MappedSuperclass
 public abstract class AbstractTranslation extends ModifiableComponent implements Translation {
 
+    @NotNull
+    @Transient  // The language must be persisted as a key annotated in the owning entity
     private Language language;
+    @NotNull
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
     private TextType type;
+    @NotEmpty
+    @Column(nullable = false)
     private String text;
+    @NotNull
+    @ManyToOne
     private AppUser createdBy;
+    @ManyToOne
     private AppUser modifiedBy;
 
     public AbstractTranslation() {
+    }
+
+    public AbstractTranslation(Language language, String text) {
+        this(language, TextType.PLAINTEXT, text);
     }
 
     public AbstractTranslation(Language language, TextType type, String text) {