+++ domain/src/main/java/ru/dpopkov/knownics/domain/TextType.java
@@ -5,5 +5,6 @@ package ru.dpopkov.knownics.domain;
  */
 public enum TextType {
     PLAINTEXT,
-    MARKDOWN
+    MARKDOWN,
+    HTML
 }