+++ domain/src/main/java/ru/dpopkov/knownics/domain/answer/SourceDetails.java
@@ -4,12 +4,24 @@ import lombok.Getter;
 import lombok.Setter;
 import ru.dpopkov.knownics.domain.ModifiableEntity;
 
+import javax.persistence.Entity;
+import javax.persistence.ManyToOne;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
+
 @Getter
 @Setter
+@Entity
 public class SourceDetails extends ModifiableEntity {
 
+    @NotNull
+    @ManyToOne
     private Source source;
+    @NotEmpty
     private String details;
+
+    public SourceDetails() {
+    }
 
     public SourceDetails(Source source, String details) {
         this.source = source;