+++ domain/src/main/java/ru/dpopkov/knownics/domain/question/Category.java
@@ -5,12 +5,14 @@ import lombok.Setter;
 import ru.dpopkov.knownics.domain.ModifiableEntity;
 
 import javax.persistence.Entity;
+import javax.validation.constraints.NotEmpty;
 
 @Getter
 @Setter
 @Entity
 public class Category extends ModifiableEntity {
 
+    @NotEmpty
     private String name;
     private String description;