+++ repository-jpa/src/test/java/ru/dpopkov/knownics/repository/jpa/ModuleTestConfiguration.java
@@ -5,5 +5,5 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 
 @SpringBootConfiguration
 @EnableAutoConfiguration
-public class DataModuleTestConfiguration {
+public class ModuleTestConfiguration {
 }