+++ history.md
@@ -1,6 +1,8 @@
 History of implementation
 -------------------------
 
+##### 2021-12-31
+* Add annotations to domain entities and embedded components. Closes #16. 
 ##### 2021-12-29
 * Add i18n for web. Closes #14. 
 * Refactor to Lombok. Closes #15.